+++ src/libam/out/FastaReadOutput.hh
@@ -25,6 +25,9 @@ public:
     void writeSE(const PairwiseAlignment& pwa) override;
     void writePE(const PairwiseAlignment& pwa1, const PairwiseAlignment& pwa2) override;
     void close() override;
+
+    bool require_alignment() const override;
+
     ~FastaReadOutput() override;
 
 private: